proj: refactor new command
